[Add] 스플래시 화면, App Coordinator 분리 및 정리
