+++ weave-iOS/Projects/DesignSystem/Resources/Icons/Icons.xcassets/Splash/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
