+++ weave-iOS/Projects/DesignSystem/Resources/Icons/Icons.xcassets/Splash/splash_logo.imageset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "images" : [
+    {
+      "filename" : "splash_logo.png",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
